zmiana miejsca skryptów
